requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
colorama>=0.4.6
requests>=2.31.0
oracledb>=1.3.2
mysql-connector-python>=8.1.0
python-multipart>=0.0.6
